improve increment and decrement docblocks
The underlying query builder method already checks if `$amount` is not
numeric and throws an exception if so. By allowing numeric-string as a
type here as well, it helps static analysis. Particularly, when casting
a decimal database value, e.g. `cost` to `decimal:2`, (which is casted
to a numeric string) we can now run
`$otherModel->increment('bounty', $model->cost)` without static analysis
erroring and requiring us to cast to a float or an int first.